Making it into a "module"
It doesn't technically make a module, but I figured it would be
easier to integrate it into the build process if it can go in the modules
folder, at least for the build.

+++ PyConnector/PyConnector.cpp
@@ -282,7 +282,7 @@ PyCode::~PyCode()
 
 Variant PyCode::evaluate(CallPoint* callPoint)
 {
-    int numParams = parqty(callPoint);
+    int numParams = (int)parqty(callPoint);
     PyGILState_STATE state;
     if (pyConnector.hasFlexSimPyController)
         state = PyGILState_Ensure();
